Configure interactions and make graphs
Allows user to select which interactions to track for paired
interactions to make the distributions. The graphs in the ComBind paper
state that MCSS only uses the RMSD of poses that actually have a MCSS.
We also expand the domain to match that of the graph, (0,15) rather than
(0,6) that is used in the paper.