No longer depend on formal charge
Salt bridge is no longer dependent on formal charge of the protein nor
ligand atoms. We check the protein for charged residues and mark all
atoms in those residues that can be charged, oxygen or nitrogen. This
way we don't have to depend on the formal charge being set for either
protein or ligand.