Basic shape in RDKit and CNN scores optional
Can now compute the shape similarity of poses using RDKits
ShapeTanimotoDist. This looks at the Tanimoto similarity of the shape of
two molecular conformations. IT does not differentiate between different
atom types, so we might need to look into ways to do that.

CNN scores are now optional for featurization, so you can just use Vina
scoring with Gnina.